# docs/server-api.openapi.yaml
openapi: 3.0.3
info:
  title: ManyRows Server-to-Server API
  version: "1.0.0"
  description: >
    The customer backend (server-to-server) API. Authenticate every request
    with a workspace API key in the `X-API-Key` header (or
    `Authorization: Bearer <key>`). Every endpoint is scoped to **members of the
    calling app**: the user pool only shares credentials across apps (SSO), it is
    not an access boundary, so a key for one app only sees and acts on users who
    have joined that app. For isolation between apps, give them separate pools.

    Requests are rate limited per API key (HTTP 429 + `Retry-After` when
    exceeded). Errors are returned as JSON `{ "error": "<code>", "message": "..." }`.
servers:
  - url: https://{host}/x/{workspaceSlug}/api/v1
    description: A workspace's server API base.
    variables:
      host:
        default: auth.example.com
        description: Your ManyRows host.
      workspaceSlug:
        default: your-workspace
        description: The workspace slug.
security:
  - ApiKeyAuth: []
...
paths:
...
  /apps/{appId}/roles:
    parameters:
      - $ref: "#/components/parameters/AppId"
    get:
      tags: [Authorization]
      summary: List roles
      description: The product's roles, each with the permission slugs it grants.
      operationId: listRoles
      responses:
        "200":
          description: Role catalog.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RolesListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Authorization]
      summary: Define a role
      description: Create a role in the product, optionally with permission slugs.
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRoleRequest" }
      responses:
        "201":
          description: The created role.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoleSummary" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: A role with that slug already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /apps/{appId}/roles/{slug}:
    parameters:
      - $ref: "#/components/parameters/AppId"
      - name: slug
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Authorization]
      summary: Get a role
      description: One role with its granted permission slugs.
      operationId: getRole
      responses:
        "200":
          description: The role.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoleSummary" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Authorization]
      summary: Update a role
      description: Update the role's name and/or its granted permission slugs (omit a field to leave it unchanged).
      operationId: updateRole
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateRoleRequest" }
      responses:
        "200":
          description: The updated role.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoleSummary" }
        "400": { $ref: "#/components/responses/BadRequest" }
    delete:
      tags: [Authorization]
      summary: Delete a role
      operationId: deleteRole
      responses:
        "204": { description: Deleted. }
        "400": { $ref: "#/components/responses/BadRequest" }
  /apps/{appId}/permissions:
    parameters:
      - $ref: "#/components/parameters/AppId"
    get:
      tags: [Authorization]
      summary: List permissions
      operationId: listPermissions
      responses:
        "200":
          description: Permission catalog.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PermissionsListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Authorization]
      summary: Define a permission
      operationId: createPermission
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePermissionRequest" }
      responses:
        "201":
          description: The created permission.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PermissionSummary" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: A permission with that slug already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /apps/{appId}/permissions/{slug}:
    parameters:
      - $ref: "#/components/parameters/AppId"
      - name: slug
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Authorization]
      summary: Get a permission
      operationId: getPermission
      responses:
        "200":
          description: The permission.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PermissionSummary" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Authorization]
      summary: Rename a permission
      operationId: updatePermission
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePermissionRequest" }
      responses:
        "200":
          description: The updated permission.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PermissionSummary" }
        "400": { $ref: "#/components/responses/BadRequest" }
    delete:
      tags: [Authorization]
      summary: Delete a permission
      operationId: deletePermission
      responses:
        "204": { description: Deleted. }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    AppId:
      name: appId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found, or the user is not a member of this app.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, description: "Stable error code, e.g. error.notFound." }
        message: { type: string, description: Human-readable message. }
      required: [error, message]
...
    RoleSummary:
      type: object
      properties:
        slug: { type: string }
        name: { type: string }
        permissions:
          type: array
          items: { type: string }
          description: Permission slugs this role grants.
    CreateRoleRequest:
      type: object
      required: [slug, name]
      properties:
        slug: { type: string }
        name: { type: string }
        permissions: { type: array, items: { type: string }, description: Permission slugs to grant. }
    UpdateRoleRequest:
      type: object
      properties:
        name: { type: string }
        permissions:
          type: array
          items: { type: string }
          description: If present, replaces the role's granted permission slugs.
    CreatePermissionRequest:
      type: object
      required: [slug, name]
      properties:
        slug: { type: string }
        name: { type: string }
    UpdatePermissionRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
    RolesListResponse:
      type: object
      properties:
        roles: { type: array, items: { $ref: "#/components/schemas/RoleSummary" } }
    PermissionSummary:
      type: object
      properties:
        slug: { type: string }
        name: { type: string }
...
    PermissionsListResponse:
      type: object
      properties:
        permissions: { type: array, items: { $ref: "#/components/schemas/PermissionSummary" } }
...